requirements.txt
openai>=1.37
tiktoken
dirtyjson
llama-cpp-python>=0.2.84
python-decouple
mistralai
